refactor: split Tip controllers, register arch allow-list, tipPins naming
