+++ development/service-ops-api/src/main/java/com/thundercrew/opsapi/dashboard/dto/DashboardMapStateResponse.java
@@ -14,7 +14,7 @@ public record DashboardMapStateResponse(
         DashboardSummary summary,
         List<BikePin> bikePins,
         List<StationPin> stationPins,
-        List<TipPin> tips
+        List<TipPin> tipPins
 ) {
     public record DashboardSummary(
             long totalBikes,
